Fix MultiProcessConsumer partitions_per_proc

+++ kafka/consumer/multiprocess.py
@@ -1,4 +1,4 @@
-from __future__ import absolute_import
+from __future__ import absolute_import, division
 
 from collections import namedtuple
 import logging
@@ -169,7 +169,7 @@ class MultiProcessConsumer(Consumer):
         # * we have an even distribution of partitions among processes
 
         if partitions_per_proc:
-            num_procs = len(partitions) / partitions_per_proc
+            num_procs = int(len(partitions) / partitions_per_proc)
             if num_procs * partitions_per_proc < len(partitions):
                 num_procs += 1
